initialize acting_seat via a hack
I was too tired to do a full migration, so calling acting_seat in
game now initializes it to 0. This is to be changed anyway once a
game starting feature is implemented and seats are randomized.

Also fixed the user_acting? method to handle cases where noone
has joined the game yet.