Broadcast all updates to game to listeners
Add turbo-stream tag to the game page to listen for changes in a
game channel. Broadcast the changes in all game state changing
controllers to the said channel.

Removed the references to @game instance variable as that errored
a lot in test while broadcasting straight from the model. It might
not be needed now in the controller broadcasting context, as I guess
the instance variable is there in every case, but I couldn't be bothered
to change the code :smoking:. It is now more explicit how the data
is passed between partials, I guess that is better.

Next step: The links to actions in the view are not rendered inside
the game partial and wont get rendered in the broadcasts. Make a
footer into the Actions card that has a link to them.